% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Calculate Cronbach's alpha on multiple items}
\usage{
alpha(items, data)
}
\arguments{
\item{items}{Vector of items names that form a scale (e.g., 5-point Likert scales)}

\item{data}{Data.frame object.}
}
\value{
A character vector.
}
\description{
Calculate Cronbach's alpha on multiple items
}
\examples{
alpha(items=c("i1","i2","i3","i4","i5"), data=cas)
}
