% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas}
\alias{cas}
\title{Patient survey data}
\format{
\subsection{\code{cas}}{

An artificial data frame with 100 rows and 5 columns:
\describe{
\item{i1 -i5}{5 survey items}
...
}
}
}
\usage{
cas
}
\description{
Artificial data of a 5 item hospital satisfaction survey to be used
for a Cronbach's alpha scale.
}
\keyword{datasets}
