% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2mcmc}
\alias{co2mcmc}
\title{Markov Chain Monte Carlo linear regression estimates of plant's CO2 uptake regressed on ambient carbon dioxide concentrations}
\format{
\subsection{\code{co2mcmc}}{

MCMC list of 3 chains with 2000 rows and 4 columns in each of the list elements:
\describe{
\item{ b0}{Regression intercept.}
\item{ b1 }{Regression coefficient of the carbon dioxide concentration effect.}
\item{ sigma}{Square root of model residual variance.}
\item{ lp__}{Total log unnormalized joint posterior density of the model parameters.}
...
}
}
}
\source{
co2mcmc is a list of 3 MCMC simulations. It estimates a linear regression using the CO2 data frame from base R.
A regression model was specified using a normal distribution maximum likelihood and normal and uniform distribution priors for
the beta coefficients and sigma parameters.
}
\usage{
co2mcmc
}
\description{
Markov Chain Monte Carlo linear regression estimates of plant's CO2 uptake regressed on ambient carbon dioxide concentrations
}
\keyword{datasets}
