% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2multi}
\alias{co2multi}
\title{Bayes class object with summarized Markov Chain Monte Carlo estimates of plant's CO2 uptake as a binary variable that is above or below the median level}
\format{
\subsection{\code{co2multi}}{

Bayes class list of 7 elements with varying rows and columns in each of the list elements:
\describe{
\item{ Posterior.Summary}{Standard element that is NA in this object because it is not applicable to this analysis.}
\item{ MCMC}{Dataframe of 4 combined MCMC chain simulations with 8000 rows and 18 columns. Column 1 is CHAIN
indicates the 4 chains, columns 2-13 are 'theta' and indicates each plant's rate of having above the median
uptake level, columns 14-17 are 'omega' group rates of 'Type' by 'Treatment', and column 18 is the overall group rate.}
\item{ Multilevel}{Posterior estimate summaries in alphabetical order of plant or group names and numerical order plant or group rates.}
\item{ Target}{Standard element that is NA in this object because it is not applicable to this analysis.}
\item{ targets}{Standard element that is NA in this object because it is not applicable to this analysis.}
\item{ R2.Summary}{Standard element that is NA in this object because it is not applicable to this analysis.}
\item{ parameter}{3 parameter names in the multilevel analysis, 'theta', 'omega', and 'omegaO'.}
...
}
}
}
\source{
co2multi is a list of summarized Markov Chain Monte Carlo estimates of plant's CO2 uptake as a binary
variable that is above or below the median level. It summarizes measurements nested within plants nested within
the 4 levels from a combination of 'Treatment' by 'Type'. An estimation used the CO2 data frame from base R.
A hierarchical estimation model was specified using a gamma distribution maximum likelihood and normal
and uniform distribution priors at various group levels.
}
\usage{
co2multi
}
\description{
Bayes class object with summarized Markov Chain Monte Carlo estimates of plant's CO2 uptake as a binary variable that is above or below the median level
}
\keyword{datasets}
