% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham.R
\name{ham}
\alias{ham}
\title{Human Augmented Matching}
\usage{
ham(source, key, ...)
}
\arguments{
\item{source}{a vector of items you want to find a match for}

\item{key}{a vector of possible matches}

\item{...}{additional arguments}
}
\value{
Nothing. A shiny app.
}
\description{
The app presents you with all possible values in \code{key} that
match a simple grep for each value of \code{source}
}
\examples{
\dontrun{
ham(source = letters, key = c(letters, paste0(letters, 1), paste0(letters, 2)))
}
}
