% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp1}
\alias{hosp1}
\title{Patient hospital intervention data, intervention group only}
\format{
\subsection{\code{hosprog}}{

An artificial data frame with 352 rows and 10 columns, intervention patients only:
\describe{
\item{survey}{Patient satisfaction survey mean score.}
\item{ los }{Hospital length of stay (los)}
\item{ cost}{Hospital stay cost}
\item{ rdm30}{Patient readmission within 30 days of discharge}
\item{ death30}{Patient death within 30 days of discharge}
\item{ female}{Patient sex, 1 indicates female, 0 otherwise}
\item{ age }{Patient age}
\item{ risk}{Patient health risk score ranging from 0 to 1}
\item{ month}{12 month indicator (1 to 12)}
\item{ program}{Indicates patient program participation. 1='yes', 0='no'}
...
}
}
}
\source{
hosp1 is a subset of the artificial dataset hosprog. It is
the intervention group's data used for single group interrupted time series.
}
\usage{
hosp1
}
\description{
Patient hospital intervention data, intervention group only
}
\keyword{datasets}
