% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infections}
\alias{infections}
\title{Hospital acquired infections (HAI) during 41 months.}
\format{
\subsection{\code{infections}}{

An artificial data frame with 41 rows and 3 columns:
\describe{
\item{ Month }{Calendar month as an integer, ranging from 1 to 41.}
\item{ HAI }{Count of hospital acquired infections (HAI) within 1 month.}
\item{ PatientDays }{Count of hospital patient days within 1 month (i.e., the sum of days in the hospital for all patients).}
...
}
}
}
\source{
infections is an artificial data frame of reasonable hospital acquired infections (HAI) estimates
made entirely for the purpose of demonstrating control charts.
}
\usage{
infections
}
\description{
Hospital acquired infections (HAI) during 41 months.
}
\keyword{datasets}
