% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsEffect.R
\name{itsEffect}
\alias{itsEffect}
\title{Interrupted time series analysis effects}
\usage{
itsEffect(model, type, interruptions)
}
\arguments{
\item{model}{an interrupted time series (ITS) model with the "lm" class,}

\item{type}{analysis type for single or multiple groups and single or multiple
time periods. If selected type="sgst", it is single-group single-time;
type="sgmt", it is single-group multiple-time; type="mgst", it is multiple-group single-time;
and type="mgmt", it is multiple-group multiple-time.}

\item{interruptions}{ITS number of interruptions as numeric vector greater than 0.}
}
\value{
a data.frame object of ITS effects and summary statistics. Row names are indicated with
first, second, and additional numbers when there are multiple interruptions (e.g., 'Intervention.2').
Generally run within the assess function.
}
\description{
Calculates effects for intervention and control groups based on interrupted
time series models from an assess class object. Within a period (or interruption),
the effect that represents the trend during the period is calculated for both
groups as well as the difference between the groups. Summary statistics
are provided that include the effect sizes, t-statistic, standard errors,
p-values, and 95\% confidence intervals of the effect sizes. These values are
provided for the intervention group, control group (when applicable), and the
differences between the two groups (Linden, 2015). These values are automatically
generated while running a model in assess and represent changes per unit of time
in an intervention period. For example, In “Summary 1”, ITS effects are the expected
differences for each time point (month, year, etc.) within that time period.
The difference between intervention and control groups is per unit of time within
the first intervention period and is analogous for subsequent periods. The summary
values are calculated only using the trend coefficients for the intervention and
control groups each value represents the change per unit of time for that period.
For example, if two interruptions in a 1-group ITS model are at months 5 and 9 of the
year with coefficients of 0.25 and 0.23, respectively, then adding the baseline trend
-0.29 (ITS.Time) and 0.25 (txp) results in a trend change of -0.04 and would be reported
under "Summary 1" for the first intervention period. Subsequently, the same coefficients
have 0.23 (txp9) added to calculate the 0.19 (= -0.29 + 0.25 + 0.23) trend change value
per unit of time in the second intervention period listed in "Summary 2".
}
\examples{
i21 <- assess(formula=survey ~ ., data=hosprog, intervention = "program",topcode =NULL,
int.time="month", regression="none", interrupt=5, its="two", newdata=TRUE, propensity=NULL)
itsEffect(model= i21$ITS, type= "mgst", interruptions= 1)

}
\references{
Linden, Ariel. (2015). Conducting Interrupted Time-series Analysis for Single- and Multiple-group
Comparisons. The Stata Journal, 2015, 15(2), 480-500, https://doi.org/10.1177/1536867X1501500208
}
