% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{losmcmc}
\alias{losmcmc}
\title{Markov Chain Monte Carlo estimates of hospital length of stay from the hosprog data frame}
\format{
\subsection{\code{losmcmc}}{

MCMC list of 4 chains with 5000 rows and 2 columns in each of the list elements:
\describe{
\item{ muOfY}{Estimate of the mean parameter.}
\item{ sigmaOfY }{Estimate of the square root of variance parameter.}
...
}
}
}
\source{
losmcmc is a list of 4 MCMC simulations. It estimates the mean and standard deviation using the artificial hosprog
data frame from ham. An estimation was specified using a normal distribution maximum likelihood and normal and uniform
distribution priors for the mean and sigma parameters.
}
\usage{
losmcmc
}
\description{
Markov Chain Monte Carlo estimates of hospital length of stay from the hosprog data frame
}
\keyword{datasets}
