% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHSN}
\alias{NHSN}
\title{NHSN standardized incidence ratios for 4 incidents, 2016-2024.}
\format{
\subsection{\code{NHSN}}{

The CDC's NHSN Acute Care Hospital data frame with 1,800 rows and 10 columns:
\describe{
\item{ Observed}{Count of observed hospital CAUTI, CLABSI, SSI, & VAE incidents}
\item{ Predicted}{Predicted number of CAUTI, CLABSI, SSI, & VAE incidents}
\item{ Incident}{Category of CAUTI, CLABSI, SSI, & VAE incidents}
\item{ State}{Indicator of location of the 50 United States}
\item{ Year}{Calendar year between 2016 - 2024}
\item{ Yr}{Ascending numeric rank of Year, ranging from 1 to 9 (2024)}
\item{ SIR}{Standardized Incidence (or Infection) Ratio}
\item{ Lower}{Lower range of the SIR's 95\% Confidence Interval}
\item{ Upper}{Upper range of the SIR's 95\% Confidence Interval}
\item{ logPredicted}{Log of Predicted, use in offset argument for Poisson regression}
...
}
}
}
\source{
The 2016-2024 NHSN SIR metrics are publicly available data and are calculated
using the 2015 national baseline and risk adjustment methodology. More information
about these metrics can be found at: cdc.gov. The incidents are from over
30,000 acute care hospitals across the USA and are for catheter-associated
urinary tract infections (CAUTI), central line-associated bloodstream infections
(CLABSI), surgical site infections (SSI), and ventilator-associated events (VAE).
}
\usage{
NHSN
}
\description{
NHSN standardized incidence ratios for 4 incidents, 2016-2024.
}
\keyword{datasets}
