% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.control.R
\name{plot.control}
\alias{plot.control}
\title{#' Shewhart control charts}
\usage{
\method{plot}{control}(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  lwd = NULL,
  col = NULL,
  iname = NULL,
  icol = "black",
  trend = FALSE,
  trcol = "gray",
  tgt = NULL,
  tgtcol = "gray",
  tpline = NULL,
  tpcol = NULL,
  cex = 1,
  cex.axis = NULL,
  cex.main = NULL,
  cex.text = NULL,
  x.axis = NULL,
  ...
)
}
\arguments{
\item{x}{control object.}

\item{y}{not currently used, default is NULL.}

\item{xlim}{specify plot's x-axis limits with a 2 element numeric vector.}

\item{ylim}{specify plot's y-axis limits with a 2 element numeric vector.}

\item{xlab}{a character vector label for the x-axis.}

\item{ylab}{a character vector label for the y-axis.}

\item{main}{the main title of the plot.}

\item{lwd}{select the line width.}

\item{col}{a 2 element character vector to specify the 1) center line and 2) both control limit colors. Defaults to, if nothing selected, c("blue", "red").}

\item{iname}{intervention name text as a single character vector. Defaults to "Intervention" if nothing is selected.}

\item{icol}{specify intervention line color as a single character vector. Defaults to "black" if nothing is selected.}

\item{trend}{a logical vector that adds an ordinary least squares trend line (i.e., simple linear regression line) when trend=TRUE. Default is FALSE.}

\item{trcol}{select a color for the trend line. Default is 'gray'.}

\item{tgt}{specify 1 or more values on the y-axis of where to add a target line when y='group'. Or 1 or more values on the y-axis of where to add a target line when y='time' or 'roll'. Default is NULL.}

\item{tgtcol}{select one or multiple colors for one or multiple target lines. Default is 'gray'.}

\item{tpline}{add one or more time point vertical line(s) using x-axis values. Default is NULL (i.e., no lines).}

\item{tpcol}{specify a color for the time point line, tpline. Default is NULL.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{cex.text}{The magnification to be used for the 'Intervention' text added into the plot relative to the current setting of 1.}

\item{x.axis}{a vector of unique character or numeric values that makes up x-axis values to
replace the intervention time variable values. This will be most helpful if you prefer current
calendar years instead of values starting at 1 (e.g., x.axis= sort(unique(data$Year)) for 1900-1999,
not 1-100). Must have equal lengths for unique x.axis values and unique replaced values.}

\item{...}{additional arguments.}
}
\value{
plot of Shewhart control charts: X-bar chart, p-chart, and u-chart.
}
\description{
Graph X-bar charts, p-charts, and u-charts. This includes
producing means for center lines, lower control limits, and upper control limits. Users can also calculate
values before and after an intervention to see if a change in the control process happened. Values are
returned in a data frame.
}
\examples{
## Hospital LOS and readmissions ##
# X-bar chart
spc_x <- control(x="los", time="month", data=hosprog, type="x", n.equal=TRUE)
# Basic X-bar chart
plot(spc_x)

# p-chart, using only the numerator (i.e., y=NULL). Specify unequal sample sizes
spc_p <- control(x="rdm30", time="month", data=hosprog, type="p", n.equal=FALSE)
# p-chart, adding target and time point lines
plot(spc_p, tgt=c(0,.25), tgtcol="green", ylim=c(0,0.4), tpline=c(4,8), tpcol= c("yellow","black"))

# u-chart for infection rates with an intervention
spc_u <- control(x="HAI", y="PatientDays", time="Month", data=infections,
type="u", n.equal=FALSE, intervention=22)
# u-chart comparing trend lines and non-default graphing options
plot(spc_u, main="u-Chart: Pre/Post Intervention", col=c("green","dodgerblue"), ylim=c(0, .0025),
trend=TRUE, trcol="red", x.axis=c((1:41+12)), xlab="Months (starting at year 2)",
icol="gray", lwd=2, cex=2, cex.axis=1.1, cex.main=1.25, cex.text=1.25)
}
