% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.alpha.R
\name{print.alpha}
\alias{print.alpha}
\title{Print alpha results}
\usage{
\method{print}{alpha}(x, ...)
}
\arguments{
\item{x}{alpha object from Cronbach's alpha calculation.}

\item{...}{Additional arguments.}
}
\value{
formatted alpha results.
}
\description{
Print alpha results
}
\examples{
print(alpha(items=c("i1","i2","i3","i4","i5"), data=cas))

}
