% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.review.R
\name{print.review}
\alias{print.review}
\title{Print Model Review Results}
\usage{
\method{print}{review}(x, digits = NULL, ...)
}
\arguments{
\item{x}{review class object of a regression model.}

\item{digits}{minimal number of significant digits. Defaults to 3.}

\item{...}{Additional arguments.}
}
\value{
formatted review results in a table.
}
\description{
Formats 'review' class results to display summary statistics. These include
the coefficient's point estimate, lower and upper confidence intervals, and
the p-value. Excludes intercept coefficient as does review().
}
\examples{
# OLS regression
print(review(assess(mpg ~ wt+hp+am, data=mtcars, regression="ols")$model) )

}
\seealso{
\code{\link[=review]{review()}} for the 'review' class object.
}
