% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Summary.R
\name{print.Summary}
\alias{print.Summary}
\title{Print Model Summary Results}
\usage{
\method{print}{Summary}(x, digits = NULL, ...)
}
\arguments{
\item{x}{Summary class object of a regression model.}

\item{digits}{minimal number of significant digits. Defaults to 3.}

\item{...}{Additional arguments.}
}
\value{
formatted Summary estimate results.
}
\description{
Formats 'Summary' class results to display summary statistics. These include
the coefficient's point estimate, lower and upper confidence intervals, and
the p-value. Excludes intercept coefficient as does Summary().
}
\examples{
# OLS regression
print(Summary(assess(mpg ~ wt+hp+am, data=mtcars, regression="ols")$model) )

}
