% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Calculate Cronbach's alpha on multiple items}
\source{
Allen, M. J., & Yen, W. M. (1979). Introduction to Measurement Theory.
Brooks/Cole.ISBN 0-8185-0283-5
}
\usage{
alpha(items, data)
}
\arguments{
\item{items}{Vector of items names that form a scale (e.g., 5-point Likert scales)}

\item{data}{Data frame object.}
}
\value{
A character vector.
}
\description{
Performs Cronbach's alpha of listed items from a data frame.
Cronbach's Alpha is a formula for estimating the internal consistency
reliability of a measurement instrument such as survey items. Survey items
can have 2 or more categories such as 5-point scales and contain 2 or more items.
}
\examples{
alpha(items=c("i1","i2","i3","i4","i5"), data=cas)
}
