% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess models with regression}
\usage{
assess(
  formula,
  data,
  regression = "none",
  did = "none",
  its = "none",
  intervention = NULL,
  int.time = NULL,
  treatment = NULL,
  interrupt = NULL,
  topcode = NULL,
  propensity = NULL,
  newdata = FALSE
)
}
\arguments{
\item{formula}{a formula object.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{regression}{Select a regression method for standard regression models
(i.e., neither DID nor ITS). Options are regression="ols" (ordinary least squares AKA linear)
or regression="logistic". Default is regression="none" for no standard regression model.}

\item{did}{option for Differences-in-Differences (DID) regression. Select did="two" for
models with only 2 time points (e.g., pre/post-test). Select did="many" for >= 3 time points
(e.g., monthly time points in 12 months of data). Default is did="none" for no DID.}

\item{its}{option for Interrupted Time Series (ITS) regression. Select its="one" for one
group (e.g., intervention only). Select its="two" for two groups (intervention and control).
Default is did="none" for no ITS.}

\item{intervention}{optional intervention variable name selected for DID, ITS, and propensity score
models that indicate which cases are in the intervention or not.}

\item{int.time}{optional intervention time variable name selected for DID or ITS models. This
indicates the duration of time relative to when the intervention started.}

\item{treatment}{optional treatment start period variable name selected for DID models.
Select 1 value from 'int.time' to indicate the start of the intervention.}

\item{interrupt}{optional interruption (or intervention) period(s) variable name selected for ITS
models. Select 1 or 2 values from 'int.time' to indicate the start and/or key intervention periods.}

\item{topcode}{optional value selected to top code Y (or left-hand side) of the formula. Analyses
will be performed using the new top coded variable.}

\item{propensity}{optional character vector of variable names to perform a propensity score model.
This requires the 'intervention' option to be selected. All models will include 'pscore' (propensity
score) in the analysis as a covariate adjustment using the propensity score.}

\item{newdata}{optional logical value that indicates if you want the new data returned. newdata=TRUE
will return the data with any new columns created from the DID, ITS, propensity score, or top coding.
The default is newdata=FALSE. No new data will be returned if none was created.}
}
\value{
a list of results from selected regression models. Will return new data if selected.
And returns relevant model information such as variable names, type of analysis, formula, and summary
of ITS effects if selected.
}
\description{
Assess models with regression
}
\examples{
summary(assess(mpg ~ hp+wt, data=mtcars, regression= "ols", topcode=30)$model)
}
\references{
Angrist, J. D., & Pischke, J. S. (2009). Mostly Harmless Econometrics:
An Empiricist's Companion. Princeton University Press. ISBN: 9780691120355.
Linden, A. (2015). Conducting Interrupted Time-series Analysis for Single- and
Multiple-group Comparisons. The Stata Journal, 15, 2, 480-500. https://doi.org/10.1177/1536867X1501500208
}
