% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{Bayes}
\alias{Bayes}
\title{Summarize Bayesian Markov Chain Monte Carlo (MCMC) object}
\usage{
Bayes(
  x,
  parameter = NULL,
  mass = NULL,
  compare = NULL,
  rope = NULL,
  newdata = FALSE
)
}
\arguments{
\item{x}{list object of MCMC chains (e.g, mcmc.list).}

\item{parameter}{single character vector name of parameter in MCMC chains to produce summary statistics. Default is NULL.}

\item{mass}{numeric vector the specifies the credible mass used in the Highest Density Interval (HDI). Default is 0.95.}

\item{compare}{numeric vector with one comparison value to determine how much of the distribution is above or below
the comparison value. Default is NULL.}

\item{rope}{numeric vector with two values that define the Region of Practical Equivalence (ROPE).
Test hypotheses by setting low and high values to determine if the Highest Density Interval (HDI)
is within or without the ROPE. Parameter value declared not credible if the entire ROPE lies
outside the HDI of the parameter’s posterior (i.e., we reject the null hypothesis). For example,
the ROPE of a coin is set to 0.45 to 0.55 but the posterior 95\% HDI is 0.61 - 0.69 so we reject
the null hypothesis value of 0.50. We can accept the null hypothesis if the entire 95\% HDI falls with the ROPE. Default is NULL.}

\item{newdata}{optional logical vector that indicates if you want the new MCMC data returned. When newdata=TRUE,
it will return the list object of MCMC chains, converted into a data frame. This new data can
be used for analysis or plots. The default is newdata=FALSE.}
}
\value{
data frame of summary statistics for MCMC parameter's distribution and/or MCMC data frame.
Statistics include highest density interval, effective sample size, proportion of distribution
within and outside of a ROPE, distribution compared with a set value, and the parameter's mean,
median, and mode. And a MCMC data frame.
}
\description{
Convert a list of Bayesian analysis chains (e.g., coda package mcmc.list objects) into a data frame
for analysis and creating plots. Calculates a set of descriptive statistics that summarize
MCMC parameters. MCMC converted to data frame and summary values are also returned as a data frame.
}
\examples{
## Hospital LOS and readmissions ##
# X-bar chart statistics
}
\references{
Kruschke, J. (2014). Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and
Stan, Second Edition. New York: Academic Press. ISBN: 9780124058880
}
