% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham.R
\name{ham}
\alias{ham}
\title{Human Augmented Matching}
\usage{
ham(source, choices, key = NULL, n = NULL, context = NULL,
  dedupe = FALSE, ...)
}
\arguments{
\item{source}{a vector of items you want to find a match for}

\item{choices}{a vector of possible matches}

\item{key}{optional ID vector to append as well to the match table}

\item{n}{maximum number of choices to present user, default is 7}

\item{context}{a data.frame with additional elements to use to decide on a
choice from among the options}

\item{dedupe}{logical, default FALSE, should only one match be selected}

\item{...}{additional arguments to runApp}
}
\value{
An object \code{match_table} to the global environment containing your
matches
}
\description{
The app presents you with all possible values in \code{key} that
match a simple grep for each value of \code{source}. If \code{context} is
provided, then the first column needs to be identical to options.
}
\examples{
\dontrun{
source_keys <- letters
choice_keys <- c(letters, paste0(letters, 1), paste0(letters, 2))
ham(source = source_keys, choices = choice_keys)
ham(source = source_keys, choices = choice_keys, key = 101:178)
context_table <- data.frame(match = choice_keys, key = 101:178, data = runif(26*3))
ham(source = source_keys, choices = choice_keys,
key = 101:178, context = context_table)
# Test duplicates
source_keys <- letters
choice_keys <- c(letters, letters, LETTERS)
context_table <- data.frame(match = choice_keys, key = 101:178,
data = runif(26*3), stringsAsFactors=FALSE)
ham(source = source_keys, choices = context_table$match, key = 101:178,
context = context_table)
}
}
