% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham.R
\name{ham}
\alias{ham}
\title{Human Augmented Matching}
\usage{
ham(source, choices, key = NULL, ...)
}
\arguments{
\item{source}{a vector of items you want to find a match for}

\item{key}{optional ID vector to append as well to the match table}

\item{...}{additional arguments}

\item{options}{a vector of possible matches}
}
\value{
An object \code{match_table} to the global environment containing your
matches
}
\description{
The app presents you with all possible values in \code{key} that
match a simple grep for each value of \code{source}
}
\examples{
\dontrun{
ham(source = letters, choices = c(letters, paste0(letters, 1), paste0(letters, 2)))
ham(source = letters, choices = c(letters, paste0(letters, 1), paste0(letters, 2)),
key = 101:126)
}
}
