% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Title}
\usage{
importance(model)
}
\arguments{
\item{model}{an assess regression model or models with lm or glm class.}
}
\value{
a data.frame object with partial X^2 summary statistics.
}
\description{
Title
}
\examples{
importance(assess(mpg ~ hp + wt, data=mtcars, regression= "ols")$model)
}
