% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Importance of variables based on partial chi-square statistic}
\usage{
importance(model)
}
\arguments{
\item{model}{an assess regression model or models with lm or glm class.}
}
\value{
a data.frame object with partial X^2 summary statistics.
}
\description{
Importance of variables based on partial chi-square statistic
}
\examples{
importance(assess(mpg ~ hp + wt, data=mtcars, regression= "ols")$model)
}
\references{
Harrell, F. E., Jr. (2016). Regression Modeling Strategies. Springer
International Publishing. ISBN: 978-3-319-19424-0.
}
