% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\title{Interpret assess model output}
\usage{
interpret(object)
}
\arguments{
\item{object}{assess class objects: alpha, ITS, DID, or ols or logistic models.}
}
\value{
a list with interpretations of alpha scales or regression model results.
}
\description{
Interpret assess model output
}
\examples{
interpret(alpha(items=c("i1","i2","i3","i4","i5"), data=cas))

hos2 <- assess(formula=survey ~ ., data=hosprog, intervention = "program",
int.time="month", treatment = 5, did="two", newdata=TRUE)
interpret(hos2)$did  #interpret(hos2) also runs, returns ITS results if present
}
