% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsEffect.R
\name{itsEffect}
\alias{itsEffect}
\title{Interrupted time series analysis effects}
\usage{
itsEffect(model, type)
}
\arguments{
\item{model}{an interrupted time series (ITS) model with the "lm" class,}

\item{type}{analysis type for single or multiple groups and single or multiple
time periods. If selected type="sgst", it is single-group single-time;
type="sgmt", it is single-group multiple-time; type="mgst", it is multiple-group single-time;
and type="mgmt", it is multiple-group multiple-time.}
}
\value{
a data.frame object of ITS effects and summary statistics.
Generally run within assess().
}
\description{
Interrupted time series analysis effects
}
\examples{
i21 <- assess(formula=survey ~ ., data=hosprog, intervention = "program",topcode =NULL,
int.time="month", regression="none", interrupt=5, its="two", newdata=TRUE, propensity=NULL)
itsEffect(model= i21$ITS, type= "mgst")
}
