% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.assess.R
\name{plot.assess}
\alias{plot.assess}
\title{Prediction plot of treatment and control groups for DID and ITS models}
\usage{
\method{plot}{assess}(
  x,
  y,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  col = NULL,
  lwd = NULL,
  cex = NULL,
  cex.axis = NULL,
  cex.lab = NULL,
  cex.main = NULL,
  arrow = FALSE,
  xshift = NULL,
  add.legend = NULL,
  ...
)
}
\arguments{
\item{x}{assess object. Either difference-in-difference or interrupted time series model with no covariate adjustment.}

\item{y}{type of model, specify either 'DID' (difference-in-difference) or 'ITS' (interrupted time series). Will not accept other models.}

\item{xlim}{specify plot's x-axis limits with a 2 value vector.}

\item{ylim}{specify plot's y-axis limits with a 2 value vector.}

\item{main}{the main title of the plot.}

\item{col}{specify intervention and control group colors in a vector. Defaults are c("blue", "red") or "blue" for single-group Interrupted Time Series models.}

\item{lwd}{select the line width.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{arrow}{logical TRUE or FALSE that indicates whether arrows and
coefficient names should be added to visualize effects. Default is FALSE.}

\item{xshift}{shifts one or some of the overlapping intervention associated arrows
along the x-axis for a better view. Vector values of length 1 or 2 can be positive
or negative. And xshift should be specified in the order of the coefficients
(e.g., "DID" before "DID.Trend" for DID models with argument did="many").}

\item{add.legend}{add a legend by selecting the location as "bottomright", "bottom", "bottomleft",
"left", "topleft", "top", "topright", "right", "center". No legend if nothing selected.}

\item{...}{additional arguments.}
}
\value{
plot of partial predictions for treatment and control groups.
}
\description{
Provides partial prediction plots for treatment and control groups from difference-in-difference (DID)
and interrupted time series (ITS) models. The graph will produce lines for treatment/intervention and
control groups to gain understanding through a visual representation of the regression coefficients.
The treatment/intervention group is represented with a blue line, the control group is represented with
a red line, and the counterfactual line, when available, is a dashed line.
}
\examples{
am2 <- assess(formula= los ~ ., data=hosprog, intervention = "program",
topcode =NULL, int.time="month", regression="none", treatment= 5,
interrupt=c(5,9), did="two", its="two", newdata=TRUE, propensity=NULL)
plot(am2, "DID", add.legend="bottomleft", ylim=c(2, 8))  #DID model
plot(am2, "ITS", add.legend="top", ylim=c(2, 8))         #ITS model
plot(am2, "DID", add.legend="topleft", main="DID study", col=c("dodgerblue","magenta"),
ylim=c(2, 8), lwd=6, cex=2, cex.axis=2, cex.lab=1.5, cex.main=3, arrow=TRUE, xshift=0.02)
}
