% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.assess.R
\name{plot.assess}
\alias{plot.assess}
\title{Prediction plot of treatment and control groups for DID and ITS models}
\usage{
\method{plot}{assess}(x, y, xlim = NULL, ylim = NULL, add.legend = NULL, ...)
}
\arguments{
\item{x}{assess object. Either difference-in-difference or interrupted time series model with no covariate adjustment.}

\item{y}{type of model, specify either 'DID' (difference-in-difference) or 'ITS' (interrupted time series). Will not accept other models.}

\item{xlim}{specify plot's x-axis limits with a 2 value vector.}

\item{ylim}{specify plot's y-axis limits with a 2 value vector.}

\item{add.legend}{add a legend by selecting the location as "bottomright", "bottom", "bottomleft",
"left", "topleft", "top", "topright", "right", "center". No legend if nothing selected.}

\item{...}{additional arguments.}
}
\value{
plot of partial predictions for treatment and control groups.
}
\description{
Provides partial prediction plots for treatment and control groups from difference-in-difference (DID)
or interrupted time series (ITS) models. It is recommended to graph for models that do not include
additional covariates (e.g., patient demographics) because the coordinates may be off. The graph
will produce lines for treatment/intervention and control groups to gain understanding
through a visual representation of the regression coefficients. The treatment/intervention
group is represented with a blue line, the control group is represented with a red line, and
the  line, when available, is a dashed line.
}
\examples{
hsur <- assess(formula=survey ~ ., data=hosprog, intervention = "program",
int.time="month", interrupt=5, its="two")
plot(x=hsur, y="ITS", ylim=c(2,5), add.legend="bottom")
}
