% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Bayes.R
\name{plot.bayes}
\alias{plot.bayes}
\title{Shewhart control charts}
\usage{
\method{plot}{bayes}(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  lwd = NULL,
  col = NULL,
  iname = NULL,
  icol = "black",
  trend = FALSE,
  trcol = "gray",
  tgt = NULL,
  tgtcol = "gray",
  tpline = NULL,
  tpcol = NULL,
  cex = 1,
  cex.lab = NULL,
  cex.axis = NULL,
  cex.main = NULL,
  cex.text = NULL,
  x.axis = NULL,
  y.axis = NULL,
  round.c = NULL,
  ...
)
}
\arguments{
\item{x}{list object of MCMC chains (e.g, mcmc.list).}

\item{y}{character vector for the type of plot to graph. Select 'post', 'check', 'multi', or 'target'
for posterior summary, posterior predictive check, multilevel or hierarchical model, or target summary plots.
Default is NULL, if nothing is selected, a posterior summary plot is produced.}

\item{xlim}{specify plot's x-axis limits with a 2 element numeric vector.}

\item{ylim}{specify plot's y-axis limits with a 2 element numeric vector.}

\item{xlab}{a character vector label for the x-axis.}

\item{ylab}{a character vector label for the y-axis.}

\item{main}{the main title of the plot.}

\item{lwd}{select the line width.}

\item{tgt}{specify 1 or more values on the y-axis of where to add one or more horizontal target lines. Default is NULL.}

\item{tgtcol}{select one or multiple colors for one or multiple target lines. Default is 'gray'.}

\item{tpline}{add one or more time point vertical lines using x-axis values. Default is NULL (i.e., no lines).}

\item{tpcol}{specify a color for the time point line, tpline. Default is NULL.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{cex.text}{The magnification to be used for the iname text added into the plot relative to the current setting of 1.}

\item{x.axis}{a vector of unique character or numeric values that makes up x-axis values to
replace the time variable values. This will be most helpful if you prefer current calendar
months/years instead of values starting at 1 (e.g., x.axis= sort(unique(data$Year)) for 1900-1999,
not 1-100). Must have equal lengths for unique x.axis values and replaced values (i.e., nrow(x)). Default is NULL.}

\item{y.axis}{a vector of unique character or numeric values that makes up y-axis values to replace
the outcome variable values. This will be most helpful if your outcome needs to be converted such as rate per
1,000 patient days (e.g., y.axis= seq(min(x$HAI)*1000, max(x$HAI)*1000, length.out=nrow(x))). Must have equal
lengths for unique y.axis values and replaced values (i.e., nrow(x)). Default is NULL.}

\item{round.c}{an integer indicating the number of decimal places when rounding numbers such as for y.axis.
Default is 2.}

\item{...}{additional arguments.}

\item{parameter}{single character vector name of parameter in MCMC chains to produce summary statistics. Default is NULL.}

\item{center}{character vector that selects the type of central tendency to use when reporting parameter values.
Choices include: 'mean', 'median', and 'mode'. Default is 'mode'.}

\item{mass}{numeric vector the specifies the credible mass used in the Highest Density Interval (HDI). Default is 0.95.}

\item{compare}{numeric vector with one comparison value to determine how much of the distribution is above or below
the comparison value. Default is NULL.}

\item{rope}{numeric vector with two values that define the Region of Practical Equivalence (ROPE).
Test hypotheses by setting low and high values to determine if the Highest Density Interval (HDI)
is within or without the ROPE. Parameter value declared not credible if the entire ROPE lies
outside the HDI of the parameter’s posterior (i.e., we reject the null hypothesis). For example,
the ROPE of a coin is set to 0.45 to 0.55 but the posterior 95\% HDI is 0.61 - 0.69 so we reject
the null hypothesis value of 0.50. We can accept the null hypothesis if the entire 95\% HDI falls with the ROPE. Default is NULL.}

\item{bcol}{a single or multiple element character vector to specify the bar color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. Defaults to, if nothing selected, 'gray'.}

\item{lcol}{a single or multiple element character vector to specify the line color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. Defaults to, if nothing selected, 'gray'.}

\item{pcol}{a single or multiple element character vector to specify the point color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. Defaults to, if nothing selected, 'gray'.}
}
\value{
plot of Shewhart control charts: X-bar charts, p-charts, and u-charts with 3-sigma control limits.
}
\description{
Graph X-bar charts, p-charts, and u-charts. This includes
producing means center lines, 3-sigma upper and lower control limits. Users can also calculate
values before and after an intervention to see if a change in the control process happened. Values are
returned in a data frame.
}
\examples{
## Hospital LOS and readmissions ##
}
\references{
Kruschke, J. (2014). Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and
Stan, Second Edition. New York: Academic Press. ISBN: 9780124058880
}
