% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Bayes.R
\name{plot.Bayes}
\alias{plot.Bayes}
\title{Bayesian plots for various analyses}
\usage{
\method{plot}{Bayes}(
  x,
  y = NULL,
  ctype = "n",
  parameter = NULL,
  center = "mode",
  mass = 0.95,
  compare = NULL,
  rope = NULL,
  data = NULL,
  dv = NULL,
  iv = NULL,
  group = NULL,
  add.data = "n",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  vlim = NULL,
  curve = FALSE,
  lwd = NULL,
  breaks = 15,
  bcol = NULL,
  lcol = NULL,
  pcol = NULL,
  xpt = NULL,
  tgt = NULL,
  tgtcol = "gray",
  tpline = NULL,
  tpcol = NULL,
  pline = 20,
  pct = 95,
  add.legend = NULL,
  legend = NULL,
  cex = 1,
  cex.lab = NULL,
  cex.axis = NULL,
  cex.main = NULL,
  cex.text = NULL,
  cex.legend = NULL,
  HDItext = 0.7,
  math = "n",
  es = "n",
  round.c = 2,
  ...
)
}
\arguments{
\item{x}{Bayes object.}

\item{y}{character vector for the type of plot to graph. Select 'post', 'dxa', 'dxd', 'dxg', 'dxt', 'check',
'multi', or 'target' for posterior summary, diagnostics (4 'dx' plots produced: autocorrelation factor,
density plots on chain convergence, Gelman-Rubin statistic, and traceplot), posterior predictive check,
multilevel or hierarchical model summary, or target summary plots. Default is 'post'.}

\item{ctype}{character vector of length == 1 that indicates posterior predictive check type when y='check'.
Posterior predictive checks allow us to see how well our estimates match the observed data. These checks are
available for Bayesian estimation of outcomes and regression polynomial trend line using various distributions in the
likelihood function. Select 'n', 'ln', 'sn', 'w', 'g', 't', 'taov', 'taov1', 'ol', 'oq','oc', 'hol', 'hoq', 'hoc',
'hlol', 'hloq', 'hloc', 'odid', 'logl', 'logq', 'logc', 'hlogl', 'hlogq', 'hlogc' for these respective options:
'Normal', 'Log-normal', 'Skew-normal', 'Weibull', 'Gamma', 't', 't: ANOVA', 't: ANOVA 1 group', 'OLS: Linear',
'OLS: Quadratic', 'OLS: Cubic', 'Hierarchical OLS: Linear', 'Hierarchical OLS: Quadratic', 'Hierarchical OLS: Cubic',
'Hierarchical Log OLS: Linear', 'Hierarchical Log OLS: Quadratic', 'Hierarchical Log OLS: Cubic', 'OLS: DID',
'Logistic: Linear', 'Logistic: Quadratic', 'Logistic: Cubic', 'Hierarchical Logistic: Linear',
'Hierarchical Logistic: Quadratic', and 'Hierarchical Logistic: Cubic'. The first 8 selections are for Bayesian
estimation of outcomes and the remaining options were developed to assess regression trend lines. Default is NULL.}

\item{parameter}{a character vector of length >= 1 or a 2 element list with the name(s) of parameter in MCMC chains to produce
summary statistics. Use a 1 element vector to get posterior estimates of a single parameter. Use a 2 or more element vector
to estimate the average joint effects of multiple parameters (e.g., average infection rate for interventions A and B when
parameter= c('IntA', 'IntB')). Use a 2 element list to perform mathematical calculations of multiple parameters (see 'math' below).
For example, use parameter=list('hospital_A', 'hospital_Z') if you want to estimate the difference between the hospital's outcomes.
Use parameter= list(c('hospital_A','hospital_B'), ('hospital_Y','hospital_Z')) to estimate how different the combined hospitals A
and B values are from the combined Hospital Y and Z values. When y='check', use either a multiple element character vector that represents
center, spread, and additional distribution parameters in order of 1st, 2nd, and 3rd distribution parameters. For example,
mean and sd for a normal distribution; mean log and sd log of a log-normal dist.; xi, omega, and alpha of a skew-normal dist.;
shape, scale, and lambda of a Weibull dist.; shape and rate of a Gamma dist.; and mean, SD and nu/d.f. of a t-distribution.
Or indicate regression parameters in order (e.g., intercept, B1, B2, etc.). When y='multi', use a multiple element character vector
to list the parameter names of the hierarchy, in order of the nesting with the lowest level first (e.g., responses, person, organization).}

\item{center}{character vector that selects the type of central tendency to use when reporting parameter values.
Choices include: 'mean', 'median', and 'mode'. Default is 'mode'.}

\item{mass}{numeric vector the specifies the credible mass used in the Highest Density Interval (HDI). Default is 0.95.}

\item{compare}{numeric vector with one comparison value to determine how much of the distribution is above or below
the comparison value. Default is NULL.}

\item{rope}{numeric vector with two values that define the Region of Practical Equivalence (ROPE).
Test hypotheses by setting low and high values to determine if the Highest Density Interval (HDI)
is within or without the ROPE. Parameter value declared not credible if the entire ROPE lies
outside the HDI of the parameter’s posterior (i.e., we reject the null hypothesis). For example,
the ROPE of a coin is set to 0.45 to 0.55 but the posterior 95\% HDI is 0.61 - 0.69 so we reject
the null hypothesis value of 0.50. We can accept the null hypothesis if the entire 95\% HDI falls with the ROPE. Default is NULL.}

\item{data}{object name for the observed data when y='check' or y='multi'.}

\item{dv}{character vector of length == 1 for the dependent variable name in the observed data frame
when y='check' or y='multi'.}

\item{iv}{character vector of length >= 1 for the independent variable name(s) in the observed data frame
when y='check' or y='multi'.}

\item{group}{character list of length == 2 for 1) the grouping variable name and 2) specific group(s) in the
observed data frame. This is primarily used for multilevel or hierarchical models when y='check' or y='multi'
that the hierarchies are based on (e.g., hospitals nested within health systems).}

\item{add.data}{character vector of length == 1 to determine the type of observed data added to the plot
when y='check'. Select 'a', 'u', 'al', 'ul', 'dg', 'n' for these options: 'All', 'Unit','All: Lines',
'Unit: Lines','DID: Groups', 'none'. Default is 'n' for none.}

\item{main}{the main title of the plot.}

\item{xlab}{a character vector label for the x-axis.}

\item{ylab}{a character vector label for the y-axis.}

\item{xlim}{specify plot's x-axis limits with a 2 element numeric vector.}

\item{ylim}{specify plot's y-axis limits with a 2 element numeric vector.}

\item{vlim}{two element vector to specify limits for minimum and maximum values used to extrapolate posterior
lines along the x-axis. For example, when drawing a log-normal distribution, we may want to have our
posterior lines fit within a narrower range while having our graph's x-axis limits extend past those lines.
If so, our value limits (vlim) help us keep our posterior predictive check lines within desired limits.
Default is NULL.}

\item{curve}{select a curve to display instead of a histogram when y='post'. Default is FALSE.}

\item{lwd}{select the line width.}

\item{breaks}{number of breaks in a histogram. Default is 15.}

\item{bcol}{a single or multiple element character vector to specify the bar or band color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. Defaults to, if nothing selected, 'gray'.}

\item{lcol}{a single or multiple element character vector to specify the line color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. When multiple lines are needed, single use lines
precede multiple use lines. For example, a single comparison value line will be assigned the first lcol
while both rope lines will be given the same color of the second lcol when y='post'. Defaults to 'gray'
if nothing selected.}

\item{pcol}{a single or multiple element character vector to specify the point color(s).
When Bayesian estimates and observed values are present, the first colors are Bayesian estimates
while the last colors are observed values. Defaults to, if nothing selected, 'gray'.}

\item{xpt}{a numeric vector of single or multiple values that indicate placement of points (+) on the
x-axis when y='check'. This is intended for the graphs with predictive checks on Bayesian estimation
(i.e., not trend lines). Default is NULL.}

\item{tgt}{specify 1 or more values on the y-axis of where to add one or more horizontal target lines. Default is NULL.}

\item{tgtcol}{select one or multiple colors for one or multiple target lines. Default is 'gray'.}

\item{tpline}{add one or more time point vertical lines using x-axis values. Default is NULL (i.e., no lines).}

\item{tpcol}{specify a color for the time point line, tpline. Default is NULL.}

\item{pline}{a numeric vector of length == 1 for the number of random posterior predictive check
lines when y='check'. Default is 20.}

\item{pct}{a numeric integer vector of length == 1 for the percentage of the posterior predictive check heavy tail lines
to be drawn when ctype= 'taov' or 'taov1'. Valid values are 0 < pct < 100. Default is 95 (e.g., 95\%).}

\item{add.legend}{add a legend by selecting the location as "bottomright", "bottom", "bottomleft",
"left", "topleft", "top", "topright", "right", "center". No legend if nothing selected.}

\item{legend}{a character vector of length >= 1 to appear when y='check' or y='multi'. Legends to represent
hierarchical estimates and observed values.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{cex.text}{The magnification to be used for the iname text added into the plot relative to the current setting of 1.}

\item{cex.legend}{The magnification to be used for the legend added into the plot relative to the current setting of 1.}

\item{HDItext}{numeric vector of length == 1 that can be a negative or positive value. Identifies placement of HDI text near credible interval
when y='post'. Values are relative to the x-axis values. Default is 0.7.}

\item{math}{mathematics function performed between multiple parameters when y='post'. Available functions are: 'add', 'subtract', 'multiply',
'divide', or 'n' for none (i,e., no functions). Indicate parameters with parameter argument.
For example, when math='subtract', use parameter=list('hospital_A', 'hospital_Z') if you want to
estimate the difference between the hospital's outcomes. Use parameter=list(c('hospital_A','hospital_B'),
('hospital_Y','hospital_Z')) to estimate how different the combined hospitals A and B values are from the
combined hospitals Y and Z. Additionally, compute statistics like the coefficient of variation when math='divide'
and parameter= list('Standard_Deviation', 'Mean'). Default is 'n' for no math function.}

\item{es}{one element vector that indicates which type likelihood distribution is relevant in calculating Jacob Cohen's
effect sizes between 2 parameters when y='post'. Options are 'beta' and 'n' for the beta distribution for binary outcomes
and none (i.e., no distribution, hence no effect size calculated). For example, to get the posterior distribution summary
for the difference between the intervention and control groups on 30-day readmissions or not, use es='beta' when y='post',
math='subtract', and parameter=list('intMean', 'ctlMean'). Default is 'n' which indicates not to calculate the effect size.}

\item{round.c}{an integer indicating the number of decimal places when rounding numbers such as for y.axis.
Default is 2.}

\item{...}{additional arguments.}
}
\value{
plot of Shewhart control charts: X-bar charts, p-charts, and u-charts with 3-sigma control limits.
}
\description{
Graph X-bar charts, p-charts, and u-charts. This includes
producing means center lines, 3-sigma upper and lower control limits. Users can also calculate
values before and after an intervention to see if a change in the control process happened. Values are
returned in a data frame.
}
\examples{
## Hospital LOS and readmissions ##
}
\references{
Kruschke, J. (2014). Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and
Stan, Second Edition. New York: Academic Press. ISBN: 9780124058880
}
