% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.group.R
\name{plot.group}
\alias{plot.group}
\title{Confidence interval graphs for group class objects}
\usage{
\method{plot}{group}(
  x,
  y = "group",
  order = "alpha",
  gcol = "blue",
  gbar = FALSE,
  pcol = "red",
  overall = FALSE,
  ocol = "gray",
  obar = FALSE,
  tgt = NULL,
  tcol = "gray",
  tpline = NULL,
  tpcol = "gray",
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  lwd = 1,
  adj.alpha = 0.4,
  cex = 1,
  cex.axis = 1,
  cex.lab = 1,
  cex.main = 1,
  round.c = 2,
  abbrv = 5,
  ...
)
}
\arguments{
\item{x}{group object.}

\item{y}{type of confidence interval object, specify either 'group', 'time', or 'roll'.}

\item{order}{specify confidence interval object order as 'alpha' or 'numeric' for alphabetical or numerical ordering.}

\item{gcol}{pick confidence interval line colors for groups. Default is 'blue'.}

\item{gbar}{logical TRUE or FALSE that indicates whether group lines have confidence bars for trend over time results. Default is FALSE.}

\item{pcol}{select point color for 'group' only confidence intervals. Default is 'red'.}

\item{overall}{logical TRUE or FALSE that indicates whether to include the overall sample confidence intervals (i.e., no grouping). Default is FALSE.}

\item{ocol}{indicate the optional overall line color. Default is 'gray' when overall=TRUE.}

\item{obar}{logical TRUE or FALSE that indicates whether to add overall confidence band. Default is FALSE.}

\item{tgt}{specify a value on the x-axis of where to add a target line. Default is NULL.}

\item{tcol}{select a color for the target line. Default is 'gray'.}

\item{tpline}{add time line (time)}

\item{tpcol}{specify a value on the y-axis of where to add a time point line. Select when y="time" or y="roll". Default is NULL.}

\item{xlim}{specify plot's x-axis limits with a 2 value vector.}

\item{ylim}{specify plot's y-axis limits with a 2 value vector.}

\item{main}{the main title of the plot.}

\item{lwd}{select the line width.}

\item{adj.alpha}{factor modifying the opacity alpha of the confidence interval bars, in the range of 0 to 1. Default is 0.4.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{round.c}{an integer indicating the number of decimal places
to be used for rounding coefficient values.}

\item{abbrv}{the minimum length of the abbreviations. Default is 5.}

\item{...}{additional arguments.}
}
\value{
plot of group level confidence intervals, including estimates over time periods.
}
\description{
Confidence interval graphs for group class objects
}
\examples{
#Simple graph for confidence intervals using the t-distribution
gr1 <- group(x="program", y="los", z="month", dataf=hosprog, dist="t",
increment=3, rolling=6)
plot(x=gr1, y="group")
#Improve the graph display using various options
plot(x=gr1, y="group", order="numeric", gcol="blue", gbar=TRUE, pcol="red",
overall=TRUE, ocol="gray", obar=TRUE, tgt=4.5, tcol="green", tpline=NULL,
tpcol="gray", xlim=c(4,5), ylim=c(.9,2.1), main=NULL, lwd=3, adj.alpha=0.2,
cex=2, cex.axis=1.25, cex.lab=1.25, cex.main=1.5, round.c=1, abbrv=5)
}
