% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.importance.R
\name{plot.importance}
\alias{plot.importance}
\title{Plot of variable importance sorted by partial chi-square statistic}
\usage{
\method{plot}{importance}(x, y, ...)
}
\arguments{
\item{x}{importance object.}

\item{y}{not currently used.}

\item{...}{additional arguments.}
}
\value{
plot of variable importance, significant variables highlighted in red.
}
\description{
Plot of variable importance sorted by partial chi-square statistic
}
\examples{
plot(importance(assess(mpg ~ hp + wt, data=mtcars, regression= "ols")$model))
}
