% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Summary.R
\name{plot.Summary}
\alias{plot.Summary}
\title{Plot of a regression model's coefficient summary}
\usage{
\method{plot}{Summary}(
  x,
  y = NULL,
  main = NULL,
  sub = NULL,
  sort = NULL,
  decreasing = NULL,
  abbrv = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  lwd = NULL,
  color = NULL,
  pcol = NULL,
  pt.cex = NULL,
  pch = NULL,
  tgt = NULL,
  tcol = NULL,
  cex = NULL,
  cex.axis = NULL,
  cex.lab = NULL,
  cex.main = NULL,
  cex.sub = NULL,
  print = NULL,
  round.c = NULL,
  ...
)
}
\arguments{
\item{x}{Summary object from assess() model or lm(), glm(), and coxph() models. For assess objects, use
Summary(my_regression$model) unless using ITS or DID models (e.g., my_did_model$DID). For model objects from
lm(), glm(), or coxph(), use Summary(my_model).}

\item{y}{not currently used.}

\item{main}{overall title for the plot, default is NULL which then lists 'Summary' for OLS regression or the type
of Summary statistic such as 'Odds Ratio' (logistic), 'Incidence Rate Ratio' (Poisson), or 'Hazard Ratio' (Cox).}

\item{sub}{a character vector for the subtitle of the plot, default is NULL which then lists the outcome name.}

\item{sort}{specify how confidence intervals are sorted. Options are 'alpha' (alphabetical coefficient names),
'coef' (coefficient values), 'p' (p-values), or 'enter' (how variables were entered in the model).
Default is 'enter'. See decreasing argument to adjust how the selection is then sorted.}

\item{decreasing}{logical TRUE or FALSE that indicates how confidence intervals are sorted, in decreasing order or not.
Defaults to FALSE.}

\item{abbrv}{the minimum length of the character length for the coefficient name abbreviations. Default is 7.}

\item{xlim}{specify plot's x-axis limits using a numeric vector with length equal to 2.}

\item{ylim}{specify plot's y-axis limits using a numeric vector with length equal to 2.}

\item{xlab}{a vector label for the x-axis. Default is a blank space. Caution: It will overlap the 'sub' argument if sub != "".}

\item{ylab}{a vector label for the y-axis. Default is a blank space.}

\item{lwd}{select the line width. Default is 1.}

\item{color}{the color to be used for the horizontal lines. Default is 'blue'.}

\item{pcol}{select the point color for the point estimates. Default is 'red'.}

\item{pt.cex}{the cex to be applied to plotting symbols, default is 1.}

\item{pch}{the plotting character or symbol to be used, default is 17.}

\item{tgt}{specify a value on the x-axis of where to add a target line to indicate when confidence intervals fall outside
of a normal range or value that indicates no difference.}

\item{tcol}{select a color for the vertical target line. Default is 'black'.}

\item{cex}{a numerical value giving the amount by which plotting text and symbols should be magnified relative to the default of 1.}

\item{cex.axis}{the magnification to be used for axis annotation relative to the current setting of cex. Default is 1.}

\item{cex.lab}{the magnification to be used for x and y labels relative to the current setting of cex. Default is 1.}

\item{cex.main}{the magnification to be used for main titles relative to the current setting of cex. Default is 1.}

\item{cex.sub}{the magnification to be used for subtitles relative to the current setting of cex. Default is 1.}

\item{print}{logical TRUE or FALSE that indicates whether to print point estimates and confidence intervals. Defaults to FALSE.}

\item{round.c}{an integer indicating the number of decimal places to be used for rounding values in the printed point estimates
and confidence intervals. Default is 4.}

\item{...}{additional arguments.}
}
\value{
plot of the Summary of regression coefficients, the display can be modified in various ways.
}
\description{
Plots a Summary class object of a regression model (Summary() with a capital 'S'). Prints a chart of
the predictor's coefficient estimates with 95\% confidence intervals (Harrell, 2015). Excludes the
intercept (if present). Horizontal line segments represent the predictor variable coefficients and
the vertical line is at 0 or 1 (e.g., odds ratio, incidence rate ratio, hazard ratio and other GLMs at 1)
to show if the proper reference is outside of the "normal range" or what represents no significant
difference. The Summary plot can be used on assess objects as well as models created from the base
package's lm() and glm() as well as coxph() from the survival package. There are various graphing
options and sorting by coefficient names, values, p-value, or the model formula.
}
\examples{
# OLS regression--most basic format with Base R lm()
m01 <- lm(mpg ~ wt+hp+am, data=mtcars)
# note it uses Summary() with a captialized 'S'
plot(x=Summary(m01))

# Using the assess function, notice 'm02$model' object below
m02 <- assess(formula=mpg ~ wt+hp+am, data=mtcars, regression="ols")
# let's see the impactful 83.5 hp increase in the 1st to 3rd quartiles
plot(x=Summary(m02$model, increase=c(hp= 83.5)))

# Logistic model
m03 <- glm(vs ~ wt+hp+am, data=mtcars, family="binomial")
# Display options, sorted plot, and printing the 95\% CIs to review
plot(x=Summary(m03, increase=c(wt=1.1375, hp= 83.5) ), color="aquamarine", lwd=4,
pt.cex= 2, tcol="magenta", sort="coef", cex.axis=2, cex.main=2, xlim=c(-.5, 3),
print=TRUE, round.c=6)

# Poisson model with an offset using assess()
m04 <- assess(formula=HAI ~  Month+ offset(log(PatientDays)),
data=infections, regression="poisson")
# Because 12 months is a meaningful period in program evaluation, Month=12
plot(x=Summary(m04$model, increase=c(Month=12)), xlim=c(0.6, 1.01),
lwd=7, color="cyan", pcol="salmon", pt.cex=2)

# NOT RUN: Cox Proportional Hazards model
#library(survival)
#m04 <- coxph(Surv(time, status) ~ age+sex+ph.karno, data=cancer)
#plot(x=Summary(m04, increase=c(age=13, ph.karno=15)))

# This also works for ITS and DID causal models through ham
im22 <- assess(formula=los ~ ., data=hosprog, intervention = "program",
int.time="month", interrupt= c(5, 9), its="two")
# The intervention group had the biggest change between the baseline and month 5
plot(Summary(im22$ITS), sort="coef", decreasing=TRUE, color="red", pcol="green")

}
\references{
Harrell, F. E., Jr. (2015). Regression Modeling Strategies, Second Edition.
Springer International Publishing. ISBN: 978-3-319-19424-0.
}
