% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.alpha.R
\name{print.alpha}
\alias{print.alpha}
\title{Print alpha results}
\usage{
\method{print}{alpha}(x, ...)
}
\arguments{
\item{x}{alpha object from Cronbach's alpha calculation.}

\item{...}{Additional arguments.}
}
\value{
formatted alpha results.
}
\description{
Formats alpha class results to display summary statistics of
scale information. These include the overall alpha, scale
mean and standard deviation, item statistics, scale statistics if
an item is removed from the scale, and the total sample size.
}
\examples{
print(alpha(items=c("i1","i2","i3","i4","i5"), data=cas))

}
