% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.interpret.R
\name{print.interpret}
\alias{print.interpret}
\title{Print interpret object}
\usage{
\method{print}{interpret}(x, ...)
}
\arguments{
\item{x}{interpret object.}

\item{...}{Additional arguments.}
}
\value{
formatted interpret object results.
}
\description{
Print interpret object
}
\examples{
print(interpret(alpha(items=c("i1","i2","i3","i4","i5"), data=cas)))
}
