% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Importance of variables based on partial chi-square statistic}
\usage{
Summary(model)
}
\arguments{
\item{model}{an assess class object or models with lm or glm class.}
}
\value{
an object with summary statistics.
}
\description{
Calculates partial chi-square (Wald chi-square for individual
coefficients) from assess class objects. The
importance is the partial chi-square minus its degrees
of freedom based on the regression coefficients (Harrell, 2015).
A higher chi-square indicates a larger effect by the predictors.
Therefore, the rank of the chi-square can indicate which predictors
can contribute more in explaining the variation in the outcome variable.
}
\examples{
# OLS regression
summary(assess(mpg ~ hp + wt + cyl, data=mtcars, regression= "ols")$model)

}
