% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Creates a Summary object of regression model results for graphing purposes}
\usage{
Summary(model, increase = NULL, coefs = NULL)
}
\arguments{
\item{model}{an assess class object or models with 'lm', 'glm', or 'coxph' class.
This includes the differences-in-differences and interrupted time series models.
For assess objects, make sure to select the correct model list element, for example,
model= my_regression$model for ordinary least squares, logistic, or Poisson models.
For 'ITS' or 'DID' models, use model= its_model$ITS or model= did_model$DID. If
using Base R's lm(), glm() or even coxph(), use the model object name (e.g., model= my_glm_model).}

\item{increase}{a named numeric vector object of the coefficient name and associated increase in the predictor.
For example, in the model y ~ a + x1 + x2, increase= c(x1= 10), will produce point estimates and confidence
intervals adjusted to a 10-unit increase in the variable x1. This is most helpful when the continuous predictor
has many levels and the interpretation of the standard 1-unit increase is less informative than a larger change
in X. For example, discussing a difference in 10 or 20 years may be more helpful in describing the impact on
30-day mortality.}

\item{coefs}{an expression defining a subset of the predictor coefficient rows (i.e., not the intercept) to view in
the Summary plot. The default is NULL, thereby using all predictor coefficients. Specify, for example, 1:2 to
view just the Summary of the first 2 coefficients.}
}
\value{
a list object with regression model summary statistics such as the point estimates, lower and upper
confidence interval limits, p-values, and additional information.
}
\description{
Produces a Summary object of regression models similar to base R's summary(). However,
the purpose is to later pass the object to the plot function to produce a summary of
only the predictor variable's coefficients (i.e.,. excludes the intercept). The main
difference between Summary() and summary() is that the ham function can create the
effects from increased predictor values (e.g., increase age by 25 years rather than
report a 1 year effect) and it has a capital letter 'S' in Summary().
}
\examples{
# OLS regression
Summary(assess(mpg ~ hp + wt + cyl, data=mtcars, regression= "ols")$model)

# Creates new classes useful for plots, e.g., plot(Summary(model))
class(Summary(assess(mpg ~ hp + wt + cyl, data=mtcars, regression= "ols")$model))

# Works with Base R
Summary(lm(mpg ~ hp + wt + cyl, data=mtcars))

# Effect after increasing hp and excluding am's coefficient
m03 <- glm(vs ~ wt+hp+am, data=mtcars, family="binomial")
plot(x=Summary(m03, increase=c(wt=1.1375, hp= 83.5), coefs= c(1, 3) ))

}
\seealso{
\code{\link[=plot.Summary]{plot.Summary()}}, \code{\link[=print.Summary]{print.Summary()}} for a plot and a formatted print output of the 'Summary' class object.
}
